+++ mygui_docking/Project1/Project1/file/ReadFile.h
@@ -43,4 +43,4 @@ void StopAndSaveDualAxisData(
 );
 void SaveLaserToXLSX(
     const std::vector<std::pair<std::chrono::system_clock::time_point,
-    std::unordered_map<uint8_t, std::vector<std::pair<std::chrono::system_clock::time_point, uint16_t>>>>>& data);
+    std::unordered_map<uint8_t, std::vector<std::pair<std::chrono::system_clock::time_point, uint32_t>>>>>& data);
